# docs/swagger.yaml
definitions:
  admin.AssignQuotaReq:
    properties:
      note:
        description: Note 备注
        type: string
      quota:
        description: Quota 分配智慧果数量
        type: integer
      user_id:
        description: UserID 用户 ID
        type: integer
      valid_period:
        description: ValidPeriod 有效期，单位小时
        type: integer
    type: object
...
  admin.UserQuotaResponse:
    properties:
      details:
        description: Details Smart Fruit User Details
        items:
          $ref: '#/definitions/repo.Quota'
        type: array
      total:
        description: Total remaining Smart Fruit
        type: integer
    type: object
  admin.UserResponse:
    properties:
      apple_uid:
        type: string
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      invite_by:
        type: integer
      invite_code:
        type: string
      phone:
        type: string
      prefer_signin_method:
        type: string
      realname:
        type: string
      status:
        type: string
      union_id:
        type: string
      updatedAt:
        type: string
      user_type:
        type: string
    type: object
  coins.ModelWithName:
    properties:
      end_at:
        type: string
      free_count:
        type: integer
      id:
        type: integer
      info:
        type: string
      model:
        type: string
      name:
        type: string
      non_cn:
        type: boolean
    type: object
...
  common.DataArray-model_ModelsDailyFree:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ModelsDailyFree'
        type: array
    type: object
...
  common.DataArray-repo_Model:
    properties:
      data:
        items:
          $ref: '#/definitions/repo.Model'
        type: array
    type: object
...
  common.DataObj-admin_UserResponse:
    properties:
      data:
        $ref: '#/definitions/admin.UserResponse'
    type: object
  common.DataObj-model_ModelsDailyFree:
    properties:
      data:
        $ref: '#/definitions/model.ModelsDailyFree'
    type: object
  common.DataObj-model_Setting:
    properties:
      data:
        $ref: '#/definitions/model.Setting'
    type: object
  common.DataObj-repo_Model:
    properties:
      data:
        $ref: '#/definitions/repo.Model'
    type: object
...
  common.EmptyResponse:
    type: object
  common.IDResponse-int64:
    properties:
      id:
        type: integer
    type: object
...
  model.ModelsDailyFree:
    properties:
      createdAt:
        type: string
      end_at:
        type: string
      free_count:
        type: integer
      id:
        type: integer
      info:
        type: string
      model_id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Setting:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      key:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
...
  repo.Model:
    properties:
      avatar_url:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/repo.ModelMeta'
      model_id:
        type: string
      name:
        type: string
      providers:
        items:
          $ref: '#/definitions/repo.ModelProvider'
        type: array
      short_name:
        type: string
      status:
        type: integer
      updatedAt:
        type: string
      version_max:
        type: string
      version_min:
        type: string
    type: object
  repo.ModelAddReq:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      meta:
        $ref: '#/definitions/repo.ModelMeta'
      model_id:
        type: string
      name:
        type: string
      providers:
        items:
          $ref: '#/definitions/repo.ModelProvider'
        type: array
      short_name:
        type: string
      status:
        type: integer
    type: object
  repo.ModelMeta:
    properties:
      input_price:
        description: InputPrice 输入 Token 价格（智慧果/1K Token），为空则与 OutputPrice 相同
        type: integer
      max_context:
        description: MaxContext 最大上下文长度
        type: integer
      output_price:
        description: OutputPrice 输出 Token 价格（智慧果/1K Token）
        type: integer
      prompt:
        description: Prompt 全局的系统提示语
        type: string
      restricted:
        description: Restricted 是否是受限制的模型
        type: boolean
      vision:
        description: Vision 是否支持视觉能力
        type: boolean
    type: object
  repo.ModelProvider:
    properties:
      id:
        description: 模型供应商 ID 优先从 channels 中查询模型供应商，不设置则根据 name 直接读取配置文件中固定的供应商配置
        type: integer
      model_rewrite:
        description: ModelRewrite 模型名称重写，如果为空，则使用模型的名称
        type: string
      name:
        description: Name 供应商名称
        type: string
    type: object
  repo.ModelUpdateReq:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      meta:
        $ref: '#/definitions/repo.ModelMeta'
      name:
        type: string
      providers:
        items:
          $ref: '#/definitions/repo.ModelProvider'
        type: array
      short_name:
        type: string
      status:
        type: integer
      version_max:
        type: string
      version_min:
        type: string
    type: object
  repo.Quota:
    properties:
      created_at:
        type: string
      expired:
        type: boolean
      id:
        type: integer
      note:
        type: string
      payment_id:
        type: string
      period_end_at:
        type: string
      period_start_at:
        type: string
      quota:
        type: integer
      rest:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: mylxsw@aicode.cc
    name: mylxsw
    url: https://ai.aicode.cc
  description: AIdea API
  title: AIdea API
  version: "1.0"
paths:
...
  /v1/admin/free-models/daily:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataArray-model_ModelsDailyFree'
      summary: Return all the free model listings.
      tags:
      - Admin:FreeModels
    post:
      consumes:
      - application/json
      parameters:
      - description: Model information
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/coins.ModelWithName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: add free model
      tags:
      - Admin:FreeModels
  /v1/admin/free-models/daily/{model_id}:
    delete:
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: Delete free model
      tags:
      - Admin:FreeModels
    get:
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataObj-model_ModelsDailyFree'
      summary: Return the specified free model information.
      tags:
      - Admin:FreeModels
    put:
      consumes:
      - application/json
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      - description: Model information
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/coins.ModelWithName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: Update free models
      tags:
      - Admin:FreeModels
...
  /v1/admin/models:
    get:
      parameters:
      - description: Sort fields, support id:desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataArray-repo_Model'
      summary: Return the list of all models.
      tags:
      - Admin:Models
    post:
      consumes:
      - application/json
      parameters:
      - description: Model information
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/repo.ModelAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.IDResponse-int64'
      summary: Add model
      tags:
      - Admin:Models
  /v1/admin/models/{model_id}:
    delete:
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: delete model
      tags:
      - Admin:Models
    get:
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataObj-repo_Model'
      summary: Retrieve detailed information for the specified model.
      tags:
      - Admin:Models
    put:
      consumes:
      - application/json
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      - description: Model information
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/repo.ModelUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: update model
      tags:
      - Admin:Models
...
  /v1/admin/quotas/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: AssignQuotaReq
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AssignQuotaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: Assign smart fruits to users
      tags:
      - Admin:Quotas
  /v1/admin/quotas/users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.UserQuotaResponse'
      summary: User Smart Fruit Details
      tags:
      - Admin:Quotas
...
  /v1/admin/settings/key/{key}:
    get:
      parameters:
      - description: Configuration item key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataObj-model_Setting'
      summary: Get the specified configuration item.
      tags:
      - Admin:Settings
  /v1/admin/settings/key/{key}/reload:
    post:
      parameters:
      - description: Configuration item key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EmptyResponse'
      summary: Reload the specified configuration item.
      tags:
      - Admin:Settings
...
  /v1/admin/users/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.DataObj-admin_UserResponse'
      summary: User details
      tags:
      - Admin:User
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"